+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [1.2.0] 2018-05-03
+## Added
+- configurable Bugsnag reporting depending on severity
+
+## [1.1.2] 2018-05-01
+- original error is reported to bugsnag instead of blunder error
+- added a stacktrace to a bugsnag report
 
 ## [1.1.1] 2018-04-23
 ### Fixed
@@ -18,7 +25,7 @@ First public release
 ### Changed
 - Exception handling no longer added around the default `MapGet` resolver middleware. This gives a dramatic performance improvement.
 ### Added
-- `Blunder.Absinthe.Dataloader.SafeKV` added to provide excpetion handling in KV Dataloaders
+- `Blunder.Absinthe.Dataloader.SafeKV` added to provide exception handling in KV Dataloaders
 ### Fixed
 - Absinthe middleware can be specified in the middleware list by a bare module name (atom). We were not handling this case in `add_error_handling`
 - Just adding blunder-absinthe to your mixfile caused runtime errors since there was no default error handler list. Making `LogError` the default.